refactor code (fix types, format, edit helpers)
