+++ src/helpers/getDirectionHelper.ts
