+++ src/helpers/getRandomHelper.ts
@@ -1,0 +1,3 @@
+export function getRandom (min: number, max: number): number {
+  return Math.floor((Math.random() * (max - min)) + min)
+}
